Refactor Container and storesMixinFactory
- Container has been renamed to Provider
- storesMixinFactory is now just appProviderMixin

With this change, any Flame app should now only receive the root
`appState` at the top entry point in the app, with all other state
being passed down as props. The `Provider` component subscribes to
state changes and provides the child component with `appState` as a
single prop which should then be destructed and passed to children.

Any child that wants to send actions should include the
`appProviderMixin` and use `context.app`.